+++ GenericModConfigMenu/Framework/ModOption/ReadOnlyModOption.cs
@@ -21,7 +21,10 @@ namespace GenericModConfigMenu.Framework.ModOption
         public override void AfterSave() { }
 
         /// <inheritdoc />
-        public override void GetLatest() { }
+        public override void BeforeMenuOpened() { }
+
+        /// <inheritdoc />
+        public override void BeforeMenuClosed() { }
 
 
         /*********