+++ GenericModConfigMenu/Framework/ModOption/SimpleModOption.cs
@@ -80,7 +80,20 @@ namespace GenericModConfigMenu.Framework.ModOption
         public override void AfterSave() { }
 
         /// <inheritdoc />
-        public override void GetLatest()
+        public override void BeforeMenuOpened()
+        {
+            this.GetLatest();
+        }
+
+        /// <inheritdoc />
+        public override void BeforeMenuClosed() { }
+
+
+        /*********
+        ** Private methods
+        *********/
+        /// <summary>Update the value from the mod configuration.</summary>
+        private void GetLatest()
         {
             this.CachedValue = this.GetValue();
         }