+++ GenericModConfigMenu/Framework/SpecificModConfigMenu.cs
@@ -80,9 +80,10 @@ namespace GenericModConfigMenu.Framework
                 string name = opt.Name();
                 string tooltip = opt.Tooltip();
 
-                opt.GetLatest();
                 if (this.InGame && opt.IsTitleScreenOnly)
                     continue;
+
+                opt.BeforeMenuOpened();
 
                 Label label = new Label
                 {
@@ -572,6 +573,9 @@ namespace GenericModConfigMenu.Framework
 
         private void Close()
         {
+            foreach (var option in this.ModConfig.ActiveDisplayPage.Options)
+                option.BeforeMenuClosed();
+
             if (this.IsSubPage)
                 this.OpenPage(null);
             else